+++ src/lib/components/filter/filter.module.ts
@@ -19,7 +19,7 @@ import { MatInputModule } from '@angular/material/input';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { ReactiveFormsModule } from '@angular/forms';
 import { L10nTranslationModule } from 'angular-l10n';
-import { TooltipModule } from '../tooltip/tooltip.module';
+import { TerraTooltipModule } from '../tooltip/terra-tooltip.module';
 
 /** Exports all the functionality needed to create a filter for a table. */
 @NgModule({
@@ -34,7 +34,7 @@ import { TooltipModule } from '../tooltip/tooltip.module';
         MatInputModule,
         MatAutocompleteModule,
         L10nTranslationModule,
-        TooltipModule
+        TerraTooltipModule
     ],
     declarations: [
         FilterActionsDirective,