+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Minor changes
 
 -   **terra-node-tree** add support for mat-icons.
--   **TooltipDirective** added its own module - **TooltipModule**.
+-   **tooltip** added `TerraTooltipModule` to be able to individually include tooltip functionality.
 
 # 12.2.0 (01.11.2021)